pdf_handler/requirements.txt
pypdf==4.3.1
PyMuPDF==1.24.9
pyyaml==6.0.2
